Revert "LAU-224 Increase session timeout. Implement IdAM token refresh. (#5)"
This reverts commit 32228a69456584c9bad93159d5b9e7bcba8cfdda.
