+++ package.json
@@ -69,12 +69,12 @@
     "nunjucks": "^3.2.2",
     "require-directory": "^2.1.1",
     "serve-favicon": "^2.5.0",
-    "ssl-config": "^1.0.0",
     "totp-generator": "^0.0.12",
     "ts-node": "^10.2.0",
     "tsconfig-paths": "^3.9.0",
     "typescript": "^4.3.5",
-    "ws": "^8.1.0"
+    "ws": "^8.1.0",
+    "ssl-config": "^1.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.8.4",
