+++ src/main/models/appRequest.ts
@@ -41,8 +41,6 @@ export interface LogData {
 
 export interface UserDetails {
   accessToken: string;
-  expiresAt: number;
-  refreshToken: string;
   idToken: string;
   id: string;
   roles: string[];
