+++ src/main/modules/oidc/index.ts
@@ -1,7 +1,28 @@
-import {AuthService, IdamGrantType} from '../../service/AuthService';
+const logger = (require('@hmcts/nodejs-logging')).Logger.getLogger('oidc');
+
 import {Application, NextFunction, Response} from 'express';
+import fetch, {Response as FetchResponse} from 'node-fetch';
 import config from 'config';
+import jwt_decode from 'jwt-decode';
 import {AppRequest} from '../../models/appRequest';
+import {HttpResponseError} from '../../util/HttpResponseError';
+
+interface IdTokenJwtPayload {
+  uid: string;
+  sub: string;
+  given_name: string;
+  family_name: string;
+  roles: string[];
+}
+
+interface IdamResponseData {
+  access_token: string;
+  refresh_token: string;
+  scope: string;
+  id_token: string;
+  token_type: string;
+  expires_in: number;
+}
 
 /**
  * Adds the oidc middleware to add oauth authentication
@@ -17,11 +38,11 @@ export class OidcMiddleware {
     '/assets/images/favicon.ico',
   ];
 
-  private authService = new AuthService();
-
   public enableFor(server: Application): void {
     const loginUrl: string = config.get('services.idam.authorizationURL');
+    const tokenUrl: string = config.get('services.idam.tokenURL');
     const clientId: string = config.get('services.idam.clientID');
+    const clientSecret: string = config.get('services.idam.clientSecret');
     const redirectUri: string = config.get('services.idam.callbackURL');
 
     server.get('/login', (req: AppRequest, res) => {
@@ -29,12 +50,56 @@ export class OidcMiddleware {
     });
 
     server.get('/oauth2/callback', async (req: AppRequest, res: Response) => {
-      this.authService.getIdAMToken(IdamGrantType.AUTH_CODE, req.session, req.query.code as string)
-        .then(() => res.redirect('/'))
-        .catch(() => res.redirect('/'));
+
+      const params = new URLSearchParams({
+        client_id: clientId,
+        client_secret: clientSecret,
+        grant_type: 'authorization_code',
+        redirect_uri: redirectUri,
+        code: req.query.code as string,
+      });
+
+      let response: FetchResponse;
+      try {
+        response = await fetch(
+          tokenUrl,
+          {
+            method: 'POST',
+            headers: {
+              'Accept': 'application/json',
+              'Content-Type': 'application/x-www-form-urlencoded',
+            },
+            body: params.toString(),
+          },
+        );
+      } catch (e) {
+        logger.error(e);
+        return res.redirect('/');
+      }
+
+      try {
+        OidcMiddleware.checkStatus(response);
+      } catch (error) {
+        logger.error(error);
+
+        const errorBody = await error.response.text();
+        logger.error(`Error body: ${errorBody}`);
+        return res.redirect('/');
+      }
+
+      const data: IdamResponseData = await response.json();
+      const jwt: IdTokenJwtPayload = jwt_decode(data.id_token);
+
+      req.session.user = {
+        accessToken: data.access_token,
+        idToken: data.id_token,
+        id: jwt.uid,
+        roles: jwt.roles,
+      };
+      req.session.save(() => res.redirect('/'));
     });
 
-    server.get('/logout', (req: AppRequest, res) => {
+    server.get('/logout', function (req: AppRequest, res) {
       req.session.user = undefined;
       req.session.save(() => res.redirect('/'));
     });
@@ -61,6 +126,15 @@ export class OidcMiddleware {
     const css = /\/main\.([a-zA-Z]*[0-9]*[a-zA-Z]*)+\.css/gm;
 
     return js.test(path) || css.test(path);
+  }
+
+  private static checkStatus(response: FetchResponse): FetchResponse {
+    if (response.ok) {
+      // response.status >= 200 && response.status < 300
+      return response;
+    } else {
+      throw new HttpResponseError(response);
+    }
   }
 
 }
