+++ src/main/modules/session/index.ts
@@ -11,7 +11,7 @@ export class SessionStorage {
   private RedisStore = ConnectRedis(session);
   private MemoryStore = require('express-session').MemoryStore;
 
-  private cookieMaxAge = 30 * (60 * 1000); // 30 minutes
+  private cookieMaxAge = 10 * (60 * 1000); // 10 minutes
 
   public enableFor(app: Application): void {
     app.use(cookieParser());
@@ -28,7 +28,7 @@ export class SessionStorage {
           secure: false,
           sameSite: 'lax',
         },
-        rolling: true, // Renew the cookie for another 30 minutes on each request
+        rolling: true, // Renew the cookie for another 20 minutes on each request
         store: this.getStore(),
       }),
     );
